==> 4_weight_checks.ipynb <==
###  4 - Weights checks

import pandas as pd
import json
import os
import io
from shutil import copyfile
import re
import itertools
from google.cloud import vision
from datetime import datetime
import numpy as np
from google.protobuf.json_format import MessageToDict
pd.set_option('display.max_rows', 10)

### Replace all weights by NA's when at least one weight is missing

df_all_mills = pd.read_csv('../Images/weight_receipt/Algo all mills.csv')
weight_col_names = ['First weight', 'Second weight', 'Net 1', 'Potongan', 'Net 2']
other_col_names = ['Mill', 'ID', 'User (middle man)', 'Date', 'Time', 'Date created', 'Time created']

df_valid = df_all_mills.copy() # dataframe where all weights are replaced by NA's if all 5 weights are not found

weights_found_list = [0]*6

def convert_weights(row):
    '''replaces all weights by NAs if all five weights are not found'''
    print('\nID', row['ID'])
    weights = [row['First weight'], row['Second weight'], row['Net 1'], row['Potongan'], row['Net 2']]
    print(weights)
    weights_not_found = sum([1 if np.isnan(w) else 0 for w in weights])
    weights_found = 5 - weights_not_found
    weights_found_list[weights_found]+=1
    if weights_not_found>0:
        print(round(weights_not_found), 'weight(s) not found, all weights replaced by NAs.')
        return [np.nan]*5+[False]
    else:
        print('All weights were found.')
        return weights+[True]

df_valid[weight_col_names+['Five weights found']] = df_valid.apply(lambda row: pd.Series(convert_weights(row)), axis=1)

df_valid

#### Summary of number of weights found

df_found_summary = pd.DataFrame(weights_found_list)
df_found_summary.columns=['Number of receipts']
df_found_summary.index = ['No weight found', 'One weight found', 'Two weights found',
                  'Three weights found', 'Four weights found', 'Five weights found']

tot_num_receipts = df_found_summary['Number of receipts'].sum()
df_found_summary.loc[len(df_found_summary)] = tot_num_receipts
df_found_summary.index = list(df_found_summary.index)[:-1] + ['Total']
df_found_summary['Percentage receipts'] = df_found_summary['Number of receipts']/tot_num_receipts
df_found_summary['Percentage receipts'] = df_found_summary['Percentage receipts'].apply(lambda x: "{0:.0f}%".format(x*100))
df_found_summary

### Check weights and try to reorder them, replace all weights by NA's if weights couldn't be checked
Check 1: 1st weight – 2nd weight = Net 1  
Check 2: Net 1 – Potongan = Net 2  
Check 3: Net 1 > Net 2  
Check 4: Potongan < .12 * Net 1

def all_perm_weights(weights):
    '''returns all permutations of a list'''
    all_perm = []
    for perm in itertools.permutations(weights):
        all_perm.append(list(perm))
    return all_perm

def reordered_weights(row):
    '''check if four equations are satisfied by trying all permutations of the found weights'''
    valid_reorder = False
    weights = list(row[weight_col_names].values)
    five_weights_found = row['Five weights found']
    if not np.isnan(weights[0]):
        ID = row['ID']
        print('\nID', ID)
        print('Checking permutations for', weights)
        weights_perms = all_perm_weights(weights)
        for i, weights_perm in enumerate(weights_perms):
            w1 = weights_perm[0]
            w2 = weights_perm[1]
            w3 = weights_perm[2]
            w4 = weights_perm[3]
            w5 = weights_perm[4]
            check_1 = w1 - w2 == w3
            check_2 = w3 - w4 == w5
            check_3 = w3 > w5
            check_4 = w4 < .12 * w3
            if check_1 and check_2 and check_3 and check_4:
                print('All checks satisfied at permutation', i)
                if i != 0:
                    valid_perm_non_zero_index.append(i)
                valid_reorder = True
                return weights_perm + [five_weights_found]
        if valid_reorder is False:
            print('Couldn\'t find a valid weight permutation. All weights will be replaced by NAs.')
            return [np.nan]*5 + [row['Five weights found']]
    return [np.nan]*5 + [row['Five weights found']]
valid_perm_non_zero_index = []
df_valid[weight_col_names+['Five weights found']] = df_valid.apply(lambda x: pd.Series(reordered_weights(x)), axis=1);

df_valid

#### Number of receipts with valid weights only after reodering
Including 19 skip receipts  
Total improvement: close to 3%

len(valid_perm_non_zero_index)

df_valid

#### Count number of receipts found and valid

df_count_receipts = pd.DataFrame()
df_count_receipts['All receipts'] = df_all_mills.groupby('Mill')['ID'].agg('count')
df_count_receipts['All weights found'] = df_valid[df_valid['Five weights found']==True].groupby('Mill')['ID'].agg('count')
df_count_receipts['All weights valid'] = df_valid[['Mill']+weight_col_names].dropna().groupby('Mill')['First weight'].agg('count')
df_count_receipts.loc[len(df_count_receipts)] = list(df_count_receipts.sum().values)
df_count_receipts.index = list(df_count_receipts.index[:-1]) + ['Total']
df_count_receipts
    
df_count_receipts['Percentage all weights found'] = df_count_receipts['All weights found']/df_count_receipts['All receipts']
df_count_receipts['Percentage all weights valid'] = df_count_receipts['All weights valid']/df_count_receipts['All receipts']
df_count_receipts['Percentage all weights valid among found'] = df_count_receipts['Percentage all weights valid']/df_count_receipts['Percentage all weights found']
df_count_receipts[['Percentage all weights found', 'Percentage all weights valid', 'Percentage all weights valid among found']] = df_count_receipts[['Percentage all weights found', 'Percentage all weights valid', 'Percentage all weights valid among found']].applymap(lambda x: "{0:.0f}%".format(x*100))

df_count_receipts

df_valid = df_valid.drop(columns=['Five weights found'])
df_valid.to_csv('../Images/weight_receipt/Algo all mills - valid images with weight checks.csv', index=False)

df_valid.sample(frac=1)

### User perpective: check image validity by checking what percentage (tau) of values were found
If more than a percentage tau of the entries are missing, the user is asked to retake the picture

tau = .8
df_all_mills = pd.read_csv('../Images/weight_receipt/Algo all mills.csv')
col_names = ['Mill', 'ID', 'User (middle man)', 'Date', 'Time',
                 'First weight', 'Second weight', 'Net 1', 'Potongan', 'Net 2']
def image_is_valid_user(ID, tau=.7):
    df = df_all_mills[df_all_mills['ID']==ID][col_names]
    if len(df) == 0:
        print('Index '+str(ID)+' not found')
        return False
    if len(df) > 1:
        print('Warning:', len(df), 'duplicate rows with same ID', ID,'Duplicates were removed')
        df = pd.DataFrame(df.iloc[0]).T
    perc_found_entries = 1-df.isnull().sum().sum()/len(col_names)
    print(round(perc_found_entries*100), '% of entries found')
    if perc_found_entries >= tau:
        return True
    else:
        print('The image is blurry, cut or invalid, please retake the picture.')
        return False

for ID in df_all_mills['ID']:
    print('\nID', ID)
    df_all_mills.loc[df_all_mills['ID']==ID, 'Valid image'] = image_is_valid_user(ID, tau=tau)

#### Summary valid receipts user perspective

df_count_valid_user = pd.DataFrame()
df_count_valid_user['Total count'] = df_all_mills.groupby('Mill')['Valid image'].agg('count')
df_count_valid_user['Valid count'] = df_all_mills[df_all_mills['Valid image']==True].groupby('Mill')['Valid image'].agg('count')
df_count_valid_user.loc[len(df_count_valid_user)] = list(df_count_valid_user.sum().values)
df_count_valid_user.index = list(df_count_valid_user.index[:-1]) + ['Total']

df_count_valid_user['Percentage valid receipts'] = df_count_valid_user['Valid count']/df_count_valid_user['Total count']
df_count_valid_user['Percentage valid receipts'] = df_count_valid_user['Percentage valid receipts'].apply(lambda x: "{0:.0f}%".format(x*100))

df_count_valid_user
