==> 5_performance_results.ipynb <==
###  5 - Performance results

import pandas as pd
import json
import os
import io
from shutil import copyfile
import re
import itertools
from google.cloud import vision
from datetime import datetime
import numpy as np
from google.protobuf.json_format import MessageToDict
pd.set_option('display.max_rows', 10)

### Compare results of algorithm with ground truth (entered manually)

def found(x):
    '''return if x is NaN or False'''
    if x:
        try:
            if np.isnan(x):
                return 0
            else:
                return 1
        except:
            return 1
    else:
        return 0
    
def to_perc(x):
    '''convert int or float to percentage'''
    try:
        return str(round(x*100))+'%'
    except:
        return x
    
def compare_algo_true(df_mill_algo, df_mill_true, columns_comp):
    '''compare dataframe built using the image recognition algorithm and true dataframe entered manually
    compare the columns columns_comp which have to be in both dataframes'''
    entry_found = df_mill_algo[columns_comp].applymap(lambda x: found(x))
    entry_found = pd.DataFrame(entry_found.apply(np.sum), columns=['Percentage of found entries'])                  
    entry_found = entry_found.apply(lambda x: x/len(df_mill_algo)).reset_index()
    
    entry_true = df_mill_algo[columns_comp] == df_mill_true[columns_comp]
    entry_true = pd.DataFrame(entry_true.apply(np.sum), columns=['Percentage of true entries'])
    entry_true = entry_true.apply(lambda x: x/len(df_mill_algo)).reset_index()

    df_res = entry_found.merge(entry_true)
    df_res['True entries among found entries'] = df_res['Percentage of true entries']/df_res['Percentage of found entries']
    df_res.set_index('index')
    df_res.loc[len(df_res)] = df_res.mean()
    df_res = pd.DataFrame(df_res).set_index('index').applymap(lambda x: to_perc(x))
    df_res.index = list(df_res.index[:-1]) + ['Average accuracy']
    df_res = df_res.reset_index()
    nbReceipts = 'n='+str(len(df_mill_true))+' receipts'
    df_res.columns = [nbReceipts] + list(df_res.columns[1:])
    df_res = df_res.set_index(nbReceipts)
    return df_res

### Save comparison in csv

def rm_sec(x):
    '''remove seconds from time'''
    try:
        return x[:-3]
    except:
        return x
    
def compare_algo_true_mill_orig(mill):
    '''compare true entries with output of detection algorithm
    before removing weights if some missing weights and before weight checks'''
    df_mill_true = pd.read_csv('../Images/weight_receipt/Reference '+mill+'.csv')
    df_mill_algo = pd.read_csv('../Images/weight_receipt/Algo '+mill+'.csv', index_col=0)
    ids = list(df_mill_true['ID'].values)
    df_mill_algo = df_mill_algo[df_mill_algo['ID'].isin(ids)]

    df_mill_algo = df_mill_algo.drop(columns=['Exit time']).rename(columns={'Entry time':'Time'})
    df_mill_algo['Time'] = df_mill_algo['Time'].apply(lambda x: rm_sec(x))
    if mill=='srjnad':
        df_mill_algo = df_mill_algo.drop(columns=['Time'])
    else:
        df_mill_true = df_mill_true.drop(columns=['Exit time']).rename(columns={'Entry time':'Time'})
        df_mill_true['Time'] = df_mill_true['Time'].apply(lambda x: rm_sec(x))
    df_mill_true['Date'] = df_mill_true['Date'].apply(lambda x: datetime.strptime(x, '%m/%d/%Y').strftime('%m/%d/%Y'))
    df_mill_true = df_mill_true.set_index('ID')
    df_mill_algo = df_mill_algo.set_index('ID')
    df_mill_algo = df_mill_algo.reindex(ids)
    columns_comp = list(df_mill_true.columns)[1:]
    return compare_algo_true(df_mill_algo, df_mill_true, columns_comp)

mills = ['arvena', 'bss', 'nhr','skip', 'srjnad']
for mill in mills:
    df_compare = compare_algo_true_mill_orig(mill)
    df_compare.to_csv('../Images/weight_receipt/Comparison '+mill+' before checks.csv')

mill = 'arvena'
compare_algo_true_mill_orig(mill)

mill = 'bss'
compare_algo_true_mill_orig(mill)

mill = 'nhr'
compare_algo_true_mill_orig(mill)

mill = 'skip'
compare_algo_true_mill_orig(mill)

mill = 'srjnad'
compare_algo_true_mill_orig(mill)

### Compare result of algorithm with ground truth (after removing weights when all weights are not found and checking the 5 equations for the weights)

def compare_algo_true_mill_post_checks(mill, df_valid):
    '''compare dataframe built using the image recognition algorithm after weight checks
    and true dataframe entered manually'''
    df_mill_algo = df_valid[df_valid['Mill']==mill].dropna(subset=['First weight'])
    df_mill_true = pd.read_csv('../Images/weight_receipt/Reference '+mill+'.csv')
    df_mill_true = df_mill_true.drop_duplicates()
    ids_df_mill_algo = list(df_mill_algo['ID'].values)
    ids_df_mill_true = list(df_mill_true['ID'].values)
    ids = [ID for ID in ids_df_mill_algo if ID in ids_df_mill_true]
    df_mill_algo = df_mill_algo[df_mill_algo['ID'].isin(ids)]
    df_mill_true = df_mill_true[df_mill_true['ID'].isin(ids)]
        
    if mill=='srjnad':
        df_mill_algo = df_mill_algo.drop(columns=['Time'])
    else:
        df_mill_true = df_mill_true.drop(columns=['Exit time']).rename(columns={'Entry time':'Time'})
        df_mill_true['Time'] = df_mill_true['Time'].apply(lambda x: rm_sec(x))
    df_mill_true['Date'] = df_mill_true['Date'].apply(lambda x: datetime.strptime(x, '%m/%d/%Y').strftime('%m/%d/%Y'))
    
    df_mill_true = df_mill_true.set_index('ID')
    df_mill_algo = df_mill_algo.set_index('ID')
    df_mill_algo = df_mill_algo.sort_index()
    df_mill_true = df_mill_true.sort_index()
    
    columns_comp = list(df_mill_true.columns)[2:]
    return compare_algo_true(df_mill_algo, df_mill_true, columns_comp)

df_valid = pd.read_csv('../Images/weight_receipt/Algo all mills - valid images with weight checks.csv')

df_valid

mill = 'arvena'
compare_algo_true_mill_post_checks(mill, df_valid)

mill = 'bss'
compare_algo_true_mill_post_checks(mill, df_valid)

mill = 'nhr'
compare_algo_true_mill_post_checks(mill, df_valid)

mill = 'skip'
compare_algo_true_mill_post_checks(mill, df_valid)

mill = 'srjnad'
compare_algo_true_mill_post_checks(mill, df_valid)
